Update Readme to include mention of jasmine 2 ajax documentation.
While attempting to use this gem, I ran into a bunch of issues due to the readme being out of date. I know the updated version is still a work in progress however I propose adding a link to the documentation on the jasmine github since it was quite useful to me.