Don't throw on uninstall if not already installed.
I've found it handy in my specs to let individual specs install jasmine-ajax as needed (sometimes after downloading local resources that are awkward to include in the source code), and use a global `afterEach` to uninstall it.  This way each spec doesn't need to remember to uinstall it itself.

Unfortunately this is not currently (before this PR) possible because: 1) `uninstall` throws if jasmine-ajax is not already installed, and 2) there's is no way (that I see) to determine ahead of time if it has been installed.

So this PR simply changes `uninstall` to not throw, since that choice seems somewhat arbitrary anyway.  Obviously I can just catch the exception in my `afterEach` if you're opposed to this change, but that's a bit ugly.