Check for a global jasmine and use it if found
This makes jasmine-ajax work with Browserify. Browserify is a CommonJS environment, but it runs in a browser and doesn't play nicely with jasmine-core (or at least I had trouble using it in my project). The way we use Jasmine - and I suspect this isn't uncommon - is we include Jasmine and our Browserify-built bundle using regular script tags. So jasmine is global and requiring-in jasmine-core is incorrect, even though we are in a CommonJS environment.

This change only checks whether a global jasmine is defined. A slightly more defensive approach - but perhaps overkill? - would be to also check for particular properties on that jasmine object. i.e.: make sure it's the right jasmine.