Add "uninstall" notes to README
This seems to be the source of some confusion (https://github.com/jasmine/jasmine-ajax/issues/123). I've been using this library for years and I even forgot to add this when I set up a new project. It's covered in the full docs but it seems worth noting for absolute beginners/people coming back to the library after some time.